vbet scraper working with geoprocessing and DGOs

+++ lib/cybercastor/cybercastor/downloadFiles.py
@@ -4,9 +4,7 @@ Demo script to download files from Data Exchange
 import sys
 import os
 import json
-import traceback
 import argparse
-import datetime
 import sqlite3
 from shutil import rmtree
 import semantic_version
@@ -20,35 +18,25 @@ igo_fields = {
     'HUC10': ogr.OFTString,
     'LevelPathI': ogr.OFTString,
     'seg_distance': ogr.OFTReal,
+    'stream_size': ogr.OFTReal,
+    # 'window_size': ogr.OFTReal,
+    # 'window_area': ogr.OFTReal,
 }
 
 dgo_fields = {
     'HUC10': ogr.OFTString,
     'LevelPathI': ogr.OFTString,
     'seg_distance': ogr.OFTReal,
-    'stream_size': ogr.OFTReal,
-    'window_size': ogr.OFTReal,
-    'centerline_length': ogr.OFTReal,
-    'window_area': ogr.OFTReal,
-    'integrated_width': ogr.OFTReal,
     'active_floodplain_area': ogr.OFTReal,
-    'active_floodplain_proportion': ogr.OFTReal,
-    'active_floodplain_itgr_width': ogr.OFTReal,
+    'active_floodplain_prop': ogr.OFTReal,
     'active_channel_area': ogr.OFTReal,
-    'active_channel_proportion': ogr.OFTReal,
-    'active_channel_itgr_width': ogr.OFTReal,
+    'active_channel_prop': ogr.OFTReal,
     'inactive_floodplain_area': ogr.OFTReal,
-    'inactive_floodplain_proportion': ogr.OFTReal,
-    'inactive_floodplain_itgr_width': ogr.OFTReal,
+    'inactive_floodplain_prop': ogr.OFTReal,
     'floodplain_area': ogr.OFTReal,
-    'floodplain_proportion': ogr.OFTReal,
-    'floodplain_itgr_width': ogr.OFTReal,
-    'vb_acreage_per_mile': ogr.OFTReal,
-    'vb_hectares_per_km': ogr.OFTReal,
-    'active_acreage_per_mile': ogr.OFTReal,
-    'active_hectares_per_km': ogr.OFTReal,
-    'inactive_acreage_per_mile': ogr.OFTReal,
-    'inactive_hectares_per_km': ogr.OFTReal
+    'floodplain_prop': ogr.OFTReal,
+    'centerline_length': ogr.OFTReal,
+    'segment_area': ogr.OFTReal,
 }
 
 huc_fields = {
@@ -125,20 +113,21 @@ def download_files(stage, hucs, us_states: str,
         try:
             # Cleanup by removing the entire project folder
             project_dir = os.path.join(os.path.dirname(output_db), project_guid)
-            rmtree(project_dir)
+            if os.path.isdir(project_dir):
+                rmtree(project_dir)
 
             # Get the inputs, intermediates and output VBET GeoPackages
             vbet_input_gpkg = download_gpkg(riverscapes_api, queries, os.path.dirname(output_db), project_guid, 'INPUTS')
             vbet_inter_gpkg = download_gpkg(riverscapes_api, queries, os.path.dirname(output_db), project_guid, 'Intermediates')
             vbet_outpt_gpkg = download_gpkg(riverscapes_api, queries, os.path.dirname(output_db), project_guid, 'VBET_OUTPUTS')
 
-            process_vbet_igos(vbet_outpt_gpkg, output_db, project_id)
-            process_vbet_dgos(vbet_inter_gpkg, output_db)
+            process_vbet_igos(vbet_outpt_gpkg, output_db, project_id, huc)
+            process_vbet_dgos(vbet_inter_gpkg, output_db, project_id, huc)
 
             for __layer_name, layer_info in lookups.items():
-                process_polygon_layer_attribute(layer_info['shapefile'], layer_info['in_field_name'], vbet_inter_gpkg, output_db)
+                process_polygon_layer_attribute(layer_info['shapefile'], layer_info['in_field_name'], layer_info['out_field_name'], vbet_inter_gpkg, output_db, huc)
 
-            process_fcodes(vbet_input_gpkg, vbet_inter_gpkg, output_db)
+            process_fcodes(vbet_input_gpkg, vbet_inter_gpkg, output_db, huc)
 
             # Cleanup by removing the entire project folder
             project_dir = os.path.join(os.path.dirname(output_db), project_guid)
@@ -160,10 +149,13 @@ def download_gpkg(riverscapes_api: str, queries: dict, top_level_dir: str, proje
     # Download the file to the local path
     actual_path = download_project_files(riverscapes_api, top_level_dir, project_guid, local_path, download_file_meta)
 
+    if os.path.isfile(actual_path) is False:
+        raise Exception(f'Unable to download file {local_path} from {download_file_meta["downloadUrl"]}')
+
     return actual_path
 
 
-def process_polygon_layer_attribute(polygon_layer: str, field_name: str, vbet_inter_gpkg: str, output_db: str) -> str:
+def process_polygon_layer_attribute(polygon_layer: str, in_field_name: str, out_field_name: str, vbet_inter_gpkg: str, output_db: str, huc: str) -> str:
     """ Determine the value of the field_name attribute for the polygon_layer that intersects the dgo_polygon
     with the largest area. Use this for determining the US state, ownership or ecoregion of the dgo_polygon"""
 
@@ -180,20 +172,19 @@ def process_polygon_layer_attribute(polygon_layer: str, field_name: str, vbet_in
     transform = osr.CoordinateTransformation(dgo_srs, poly_srs)
 
     # Loop over the DGOs, transform the polygon and find the largest intersection
-    dgo_results = {}
+    dgo_results = []
     for dgo_feature in dgo_layer:
-        clone_dgo = dgo_feature.Clone()
+        clone_dgo = dgo_feature.GetGeometryRef().Clone()
         clone_dgo.Transform(transform)
-        dgo_huc = clone_dgo.GetField('HUC10')
-        dgo_level_path = clone_dgo.GetField('LevelPathI')
-        dgo_seg_distance = clone_dgo.GetField('seg_distance')
+        dgo_level_path = dgo_feature.GetField('LevelPathI')
+        dgo_seg_distance = dgo_feature.GetField('seg_distance')
 
         # Use a spatial attribute filter to get just the polygons that intersect the DGO
         intersection_areas = {}
         poly_layer.SetSpatialFilter(clone_dgo)
         for poly_feature in poly_layer:
             geom = poly_feature.GetGeometryRef()
-            output_value = poly_feature.GetField(field_name)
+            output_value = poly_feature.GetField(in_field_name)
 
             intersection_result = geom.Intersection(clone_dgo)
             if intersection_result is not None:
@@ -204,16 +195,17 @@ def process_polygon_layer_attribute(polygon_layer: str, field_name: str, vbet_in
 
         if len(intersection_areas) > 0:
             max_area = max(intersection_areas, key=lambda k: intersection_areas[k])
-            dgo_results.append((dgo_huc, dgo_level_path, dgo_seg_distance, max_area))
+            dgo_results.append((max_area, huc, dgo_level_path, dgo_seg_distance))
 
     # Update the DGO table with the results
-    with sqlite3.connect(output_db) as conn:
-        curs = conn.cursor()
-        curs.executemany(f'UPDATE vbet_dgos SET {field_name} = ? WHERE HUC10 = ? AND LevelPathI = ? AND seg_distance = ?', dgo_results)
-        conn.commit()
+    if len(dgo_results) > 0:
+        with sqlite3.connect(output_db) as conn:
+            curs = conn.cursor()
+            curs.executemany(f'UPDATE vbet_dgos SET {out_field_name} = ? WHERE HUC10 = ? AND LevelPathI = ? AND seg_distance = ?', dgo_results)
+            conn.commit()
 
 
-def process_fcodes(vbet_input_gpkg: str, vbet_inter_gpkg: str, output_gpkg: str) -> None:
+def process_fcodes(vbet_input_gpkg: str, vbet_inter_gpkg: str, output_gpkg: str, huc: str) -> None:
 
     # DGO polygon layer
     dgo_dataset = ogr.Open(vbet_inter_gpkg)
@@ -228,13 +220,12 @@ def process_fcodes(vbet_input_gpkg: str, vbet_inter_gpkg: str, output_gpkg: str)
     transform = osr.CoordinateTransformation(dgo_srs, poly_srs)
 
     # Loop over the DGOs, transform the polygon and find the largest intersection
-    dgo_results = {}
+    dgo_results = []
     for dgo_feature in dgo_layer:
-        clone_dgo = dgo_feature.Clone()
+        clone_dgo = dgo_feature.GetGeometryRef().Clone()
         clone_dgo.Transform(transform)
-        dgo_huc = clone_dgo.GetField('HUC10')
-        dgo_level_path = clone_dgo.GetField('LevelPathI')
-        dgo_seg_distance = clone_dgo.GetField('seg_distance')
+        dgo_level_path = dgo_feature.GetField('LevelPathI')
+        dgo_seg_distance = dgo_feature.GetField('seg_distance')
 
         # Use a spatial attribute filter to get just the polygons that intersect the DGO
         intersection_areas = {}
@@ -252,13 +243,14 @@ def process_fcodes(vbet_input_gpkg: str, vbet_inter_gpkg: str, output_gpkg: str)
 
         if len(intersection_areas) > 0:
             max_length = max(intersection_areas, key=lambda k: intersection_areas[k])
-            dgo_results.append((dgo_huc, dgo_level_path, dgo_seg_distance, max_length))
+            dgo_results.append((max_length, huc, dgo_level_path, dgo_seg_distance))
 
     # Update the DGO table with the results
-    with sqlite3.connect(output_gpkg) as conn:
-        curs = conn.cursor()
-        curs.executemany('UPDATE vbet_dgos SET FCode = ? WHERE HUC10 = ? AND LevelPathI = ? AND seg_distance = ?', dgo_results)
-        conn.commit()
+    if len(dgo_results) > 0:
+        with sqlite3.connect(output_gpkg) as conn:
+            curs = conn.cursor()
+            curs.executemany('UPDATE vbet_dgos SET FCode = ? WHERE HUC10 = ? AND LevelPathI = ? AND seg_distance = ?', dgo_results)
+            conn.commit()
 
 
 def get_project_model_version(project_info: Dict) -> str:
@@ -349,27 +341,32 @@ def upsert_project(output_gpkg: str, project_guid: str, project_type: str, huc: 
     return project_id
 
 
-def process_vbet_igos(vbet_output_gpkg, scrape_output_gpkg, project_id: int) -> None:
+def process_vbet_igos(vbet_output_gpkg, scrape_output_gpkg, project_id: int, huc10: str) -> None:
     """Note project ID is the SQLite integer ID, not the GUID string"""
 
     with GeopackageLayer(vbet_output_gpkg, layer_name='vbet_igos') as input_lyr:
         with GeopackageLayer(scrape_output_gpkg, layer_name='vbet_igos', write=True) as output_lyr:
 
             for in_feature, _counter, _progbar in input_lyr.iterate_features():
-                field_values = {field: in_feature.GetField(field) for field in dgo_fields}
+                field_values = {field: in_feature.GetField(field) for field in igo_fields if field != 'HUC10'}
                 field_values['project_id'] = project_id
-
+                field_values['HUC10'] = huc10
                 output_lyr.create_feature(in_feature.GetGeometryRef(), field_values)
 
 
-def process_vbet_dgos(vbet_intermediate_gpkg: str, scrape_output_gpkg: str) -> None:
+def process_vbet_dgos(vbet_intermediate_gpkg: str, scrape_output_gpkg: str, project_id: int, huc10: str) -> None:
 
     with sqlite3.connect(scrape_output_gpkg) as conn:
         curs = conn.cursor()
         with GeopackageLayer(vbet_intermediate_gpkg, layer_name='vbet_dgos') as input_lyr:
             for in_feature, _counter, _progbar in input_lyr.iterate_features():
 
-                field_key_values = {field: in_feature.GetField(field) for field in dgo_fields}
+                field_key_values = {field: in_feature.GetField(field) for field in dgo_fields if field != 'HUC10'}
+                field_key_values['HUC10'] = huc10
+
+                if field_key_values['seg_distance'] is None:
+                    continue
+
                 field_names = list(field_key_values.keys())
                 curs.execute(f'INSERT INTO vbet_dgos ({", ".join(field_names)}) VALUES ({", ".join(["?"] * len(field_names))})', [field_key_values[field] for field in field_names])
 
@@ -493,16 +490,11 @@ def build_output(output_gpkg: str, huc10_path: str, hucs: list, huc_attributes: 
     with GeopackageLayer(output_gpkg, layer_name='vbet_igos', delete_dataset=True) as out_lyr:
         out_lyr.create_layer(ogr.wkbPoint, epsg=4326, fields=vbet_igo_fields)
 
-    with GeopackageLayer(output_gpkg, layer_name='hucs', delete_dataset=True) as out_lyr:
+    with GeopackageLayer(output_gpkg, layer_name='hucs', write=True) as out_lyr:
         out_lyr.create_layer(ogr.wkbMultiPolygon, epsg=4326, fields=huc_fields)
-
-    # Load the HUC polygons
-    with ShapefileLayer(huc10_path) as huc_lyr:
-        with GeopackageLayer(output_gpkg, layer_name='hucs', write=True) as out_lyr:
+        with ShapefileLayer(huc10_path) as huc_lyr:
             for in_feature, _counter, _progbar in huc_lyr.iterate_features():
                 out_lyr.create_feature(in_feature.GetGeometryRef(), {field: in_feature.GetField(field) for field in ['huc10']})
-
-    out_lyr = None
 
     # Create database tables
     with sqlite3.connect(output_gpkg) as conn:
@@ -518,21 +510,17 @@ def build_output(output_gpkg: str, huc10_path: str, hucs: list, huc_attributes: 
 
         conn.execute(f'CREATE TABLE vbet_dgos ({",".join(field_list)})')
         conn.execute('ALTER TABLE vbet_dgos ADD COLUMN project_id INTEGER')
-        conn.execute('CREATE UNIQUE INDEX pk_vbet_dgos ON vbet_dgos (HUC10, LevelPathI, seg_distance)')
+        conn.execute('CREATE INDEX pk_vbet_dgos ON vbet_dgos (HUC10, LevelPathI, seg_distance)')
         conn.execute('CREATE INDEX vbet_dgos_project_id_idx ON vbet_dgos (project_id)')
 
         # Add the polygon lookup fields (US state, ecoregion, ownership) to the DGO table
         for __lookup_name, lookup_info in lookups.items():
             conn.execute(f'ALTER TABLE vbet_dgos ADD COLUMN {lookup_info["out_field_name"]} TEXT')
 
-        # Incorporate the HUC attribute columns
-        igo_field_list = []
-        for field_name, d_type in igo_fields.items():
-            field_type = 'TEXT' if d_type == ogr.OFTString else 'REAL'
-            igo_field_list.append(f'{field_name} {field_type}')
-        conn.execute(f'CREATE TABLE vbet_igos ({",".join(igo_field_list)})')
-        conn.execute('ALTER TABLE vbet_igos ADD COLUMN project_id INTEGER')
+        conn.execute('ALTER TABLE vbet_dgos ADD COLUMN FCode TEXT')
+
         conn.execute('CREATE INDEX vbet_igos_project_id_idx ON vbet_igos (project_id)')
+        conn.execute('CREATE INDEX vbet_igos_huc10_idx ON vbet_igos (huc10, levelPathI, seg_distance)')
 
         # Create table to track projects
         conn.execute('CREATE TABLE projects (id INTEGER PRIMARY KEY, guid TEXT NOT NULL, huc10 TEXT, project_type TEXT, model_version TEXT, status TEXT, metadata TEXT)')
@@ -541,6 +529,8 @@ def build_output(output_gpkg: str, huc10_path: str, hucs: list, huc_attributes: 
 
         # Triggers need to be off for the updates to feature class
         triggers = drop_triggers(conn)
+
+        conn.execute('CREATE INDEX ix_hucs_huc10 ON hucs (huc10)')
 
         # Enrich the HUCs with attributes from the JSON file
         with open(huc_attributes) as fhuc:
